Add support for documenting values of constants.
A common pattern in CRD APIs is to use enum validation and a
corresponding Go type that has constant values declared for
each enumerated value.

This change detects constants that are declared as instances of
an underlying type, and allows the type template to format the
values and comments in to a table.

There are 3 parts to the change:

1. Capture the constant types in a new field on the `apiPackage`.
2. Export the constant value through the `typeDisplayName` helper.
3. Export the constants of a given type through `constantsOfType`.

Signed-off-by: James Peach <jpeach@apache.org>